refactor(settings): relocate SettingsAdminContent component
Moved SettingsAdminContent to a more appropriate module directory for improved project structure and maintainability. Updated all import paths to reflect the new location.
